Bullets disappear locally when they hit another player.
If a bullet collides with a player, the bullet's mesh renderer is turned off, effectively turning it invisible. Invisible bullets cannot cause a player to lose lives.
Bullet visbility is not synchronized through the server, but is determined locally. This will not likely be an issue since bullets move quite quickly, and player health is still syncrhonized through teh esrver.